I Restructured Redux Slice Designs to Optimize Application Performance
1. Earlier in tweetSlice I was using 3 differnt
arrays for 3 different types of tweet display
outputs. One array I was using for tweets of user
whose profile is being viewed, one I was using for
all exisiting tweets which have to be displayed on
home feed in "For you" section, and the 3rd array
I was using for displaying tweets of people who
are being followed by the user who is logged in
right now

---> But I realised that I don't need 3 arrays in
tweetSlice, no matter what type of tweets are
being displayed, at the end only 1 type of tweets
are displayed on the screen at a time.

---> By making this observation I realised that
I can optimize the frontend system design on this
basis. I can just have one displayTweets array in
tweetSlice and then I can keep calling different
api endpoints as per need and then set data just
into that single array because there is no need
of displaying 2 types of tweets at the same time

2. I also made tweet likes to happen in real time
without any flickers in UI.

---> Earlier I was using local useState hooks in
components and I was dependent on backend response.
Based on the backend response I was making frontend
refreshes which was causing a flicker effect to
occur when a lot of tweets were liked at a fast pace

---> Now I have completely switched to redux based
UI updates only and the UI is working really smooth
because of that. Right now the UI is not dependent on
backend response until and unless it is an error which
is occuring and like this state management in frontend
is working good